[TASK] Update nodejs container image versions
New core-testing-* images has been build and published in
the TYPO3 testing-infrastructure repository and pulled to
the core ci gitlab runner instances. [1][2]

This change modifies `Build/Scripts/runTests.sh` to set
new image version to ensure the usage for all changes
based on this change.

For information what changed in the image see the patches
in the testing-infrastrucure repository linked below.

* core-testing-nodejs18 => 1.4
* core-testing-nodejs18-chrome => 1.4

[1] https://nodejs.org/en/blog/vulnerability/april-2024-security-releases/
[2] https://git.typo3.org/typo3/CI/testing-infrastructure/-/commit/712e35c72cd7788d1d6311714b5a0dfe46e743a7

Resolves: #103518
Releases: main, 12.4
Change-Id: Ibaafbd465a21d631f6726064255241aa4b42e726
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83652
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>