[BUGFIX] Fix wrong GUI crop area placement on edge cases
In the image editor, the "Cropped area" was configured with a `ceil`ed
width and height. Since all other dimensions are `floor`ed, this might
cause an overlap.
Due to the configured cropping mode, defining that the cropping area
must never exceed the image dimensions, the cropper library places the
cropper to a position that's mathematically correct but not desired.

To solve this issue, the `width` and `height` properties fed to the
cropper are now `floor`ed as well, basically aligning with the rest of
the coordinates.

Resolves: #101520
Releases: main, 12.4, 11.5
Change-Id: I981bd185eaf0db60662a5a6b7b423d84112c2150
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83659
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>