[BUGFIX] Prevent indexed_search crash in BE
'phash' has been changed from an int to a string
with #102975. This leads to a BE crash when selecting
details of an indexed page.

Make argument 'pageHash' a string instead of an int
within the extbase based BE module controller
statisticDetailsAction() to enable proper extbase
mapping of that argument. Also make this argument
non-optional and deal with the value as string in
the queries.

Resolves: #103541
Related: #102975
Releases: main
Change-Id: I14632eab767fcbe861db9e3bafa520b0d16b56a5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83675
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>