[TASK] Add `npm` command dispatcher to `Build/Scripts/runTests.sh`
`Build/Scripts/runTests.sh` is the monorepo main dispatcher of
script and tools to ensure same working state across developer
host system and installation by using well defined containers.
This mitigates issues with missing dependencies or wrong version
of tools on develper machines.

Until now local installation of `node` and `npm` in the matching
version was mandatory to work with the node ecosystem like manage
dependencies, executing node based tools or the build chain for
CSS and TypeScript/JavaScript.

This change adds now a `npm` dispatcher to `runTests.sh` similar
to the `composer` command dispatcher introduced with #103045 to
allow executing npm commands and defined scripts in `package.json`
in a well-defined environment and removes the hard-requirement of
installing it locally for new contributors.

Added `runTests.sh` suite:

* ./Build/Scripts/runTests.sh -s npm --  <npm-command-arguments>

which allows now to using npm commands always with the correct
npm and node version defined for the mono-repository state.

Some examples:

  ./Build/Scripts/runTests.sh -s npm -- audit
  ./Build/Scripts/runTests.sh -s npm -- ci
  ./Build/Scripts/runTests.sh -s npm -- run build
  ./Build/Scripts/runTests.sh -s npm -- run watch:build
  ./Build/Scripts/runTests.sh -s npm -- install \
    --save bootstrap@^5.3.2

Resolves: #103552
Related: #103045
Releases: main, 12.4, 11.5
Change-Id: Ie9de71f24dfd4805a98be7dc1b1604e5d0146eec
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83690
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Thomas Hohn <tho@gyldendal.dk>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>