[TASK] Avoid indexed_search EXTCONF addRootLineFields
Did you know about
$GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['indexed_search']['addRootLineFields']?

No? Well, that's probably because *nobody* does! According to
a TER search, it has never been used.

Why? Well, it's an undocumented feature in a detail functionality
of indexed_search, and last traces at more obvious places have
been removed in v6.0.

To use it, a developer not only needs to understand the details
of 'rootline_uids' in indexed_search, it also needs to find that
"hook", grasp it, add a database field in table index_section,
and add FE rendering to deal with it.

The patch removes this code without substitution: It has never
been used in the wild, is not documented, is most likely broken
if anyone actually tries to use it, there has never been a
single issue in forge related to it, it is embedded in an area that
is hard to understand in the first place, it could be resolved
using an XCLASS more easily, and its naming collides with the
well-known FE addRootLineFields, which is of course a totally
different thing.

Resolves: #103553
Releases: main
Change-Id: I311e85e207cecc6457f66d0e3f26b7f4e256ddd3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83691
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>