[TASK] Remove `HtmlParser->caseShift()`
The method `HtmlParser->caseShift()` is used to normalize the casing of
values that are about to get compared, depending on whether a
case-sensitive comparison is wanted at all which isn't, by default [1].

Also, the method had a runtime cache that doesn't have a great benefit.
Therefore, the method is removed and normalization is done at the former
method calls.

[1] https://docs.typo3.org/m/typo3/reference-typoscript/main/en-us/Functions/HtmlparserTags.html#fixattrib-attribute-casesensitivecomp

Resolves: #103565
Releases: main
Change-Id: I260daeaa677418e5f7a619176e1fae17e5f68bb2
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83704
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>