[BUGFIX] Correctly infer site for new page
SiteResolving will now correctly interpret a temporary ID of a new page (e.g., "NEW12345678ABCDEF") as an invalid ID and try to infer the site using `$result['effectivePid']`.

This makes it possible to access site configuration properties within the TCA also on pages that have yet to be created.

Resolves: #103573
Releases: main, 12.4, 11.5
Change-Id: I174e71e084e96b24ebe721ad0336a3dd1a8115d9
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85457
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>