[BUGFIX] Avoid "Save changes and view" in FormEngine if form is invalid
The implementation of the "Save changes and view" modal triggered in
FormEngine via the "View" button is prone to data inconsistency as it
may save the form and bypass the FormEngine validation.

The "View" modal now checks whether the current FormEngine state is
invalid and therefore won't render the "Save changes and view" button.

Resolves: #103637
Resolves: #98333
Releases: main, 13.4, 12.4
Change-Id: I6cbe8fdc5d56e5c616da07c84f12c1e4c1bbfd48
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83838
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>