[BUGFIX] Guard nullable getNormalizedParams in getIpFunction
In case of CLI context, normalized params are not available and as they
are nullable, a guard condition is required before accessing methods.

Note: The access via request object was added in #100047 along with a
deprecation of using the ip() function in TSConfig. However, there is a
special case, where plain TypoScript is loaded even in CLI context. For
example this can happen when EXT:form listens to FlexForm data structure
parsing event. Then it needs information from TypoScript on how to
manipulate FlexForm based on YAML definitions.

Resolves: #103644
Related: #100047
Releases: main, 12.4
Change-Id: Id2d166fd11794db9b7a8102677e5b9305fe18c59
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83842
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Nikita Hovratov <nikita.h@live.de>