[BUGFIX] Better refindex for MM and multi workspaces
Calculation of reference index for workspace relations
has a nasty bug in getListOfActiveWorkspaces(): It
always returns only the first workspace id when
there are multiple.

This isn't good in v12 since it creates broken or
missing index entries with MM. It is worse in v13,
since refactorings of the class use the method
for some additional operations.

The patch fixes the broken helper method.

Resolves: #103655
Related: #96067
Related: #103661
Releases: main, 12.4
Change-Id: I989b22f9c25b22f52030c3dbce742074d2e09f97
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83920
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>