[BUGFIX] Correctly purge RelationHandler readMM() relations
Resolving relations in RelationHandler for MM in
workspaces is a two fold process: readMM() first
finds all relation "candidates". A second step
then filters live MM relations from the candidate
list that have a workspace overlay.

In general, it *should* be possible to combine this
rather expensive operation into a single query. This
however requires a bigger rewrite of the class,
which absoletly needs to happen at some point,
especially for performance considerations.

There is however a significant bug the patch aims
to solve now: The result set of MM relations tends
to return relations not only of the workspace-uid
that has been requested, but relations of other
workspaces as well, due to an insufficient query
with improper filter logic in the second part of
the process.

This patch is important for a bug fix within
ReferenceIndex, which - when fixed - triggers
the currently broken RelationHandler relation
calculation.

Resolves: #103661
Related: #103655
Releases: main, 12.4
Change-Id: If42c138321f0043bf9d8c363f0e20e61d649e1e1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83917
Reviewed-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>