[TASK] Use "new record wizard" to render dashboard widgets
In order to streamline the backend UI and to re-use introduced web
components, the "Add widget" wizard in the dashboard module now utilizes
the `@typo3/backend/new-record-wizard` module that is already known from
the Content Element creation wizard.

As a small quality-of-life change, the modal title now includes the name
of the dashboard as well.

Resolves: #103697
Releases: main
Change-Id: I58f069e5d3f57484cba8e020f51ab82cf923d25a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83937
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>