[TASK] Remove usage of "newer" in grunt build
The grunt module `grunt-newer` was introduced with #90200 to help
improving build times on consecutive single builds by running tasks on
changed files only.

This is nowadays not required anymore in JS build context, as it's
possible in TypeScript to compile changed files only, which is already
incorporated for quite a while. Therefore, the usage of `newer` is now
removed in this context. However, to avoid running `terser` multiple
times on the same, previously compiled files, the directory
`Build/JavaScript` now gets removed before each TypeScript compilation.

The postcss calculation is fast enough to not require grunt-newer at
all, therefore grunt-newer is dropped here as well.

Executed commands:

    npm remove grunt-newer

Resolves: #103712
Releases: main, 12.4
Change-Id: I1c6b88945a45cb9a394a2beaa2ec8198865880b0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84006
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>