[TASK] Use PAGEVIEW in EXT:styleguide site set
The EXT:styleguide extension currently
provides two ways of frontend rendering.
The "old" way using a `sys_template`
record and the "new" way, using site sets.

To not only test the new site sets feature,
but also the new PAGEVIEW cObject is the
corresponding TypoScript now using the
PAGEVIEW for page rendering.

This one hand is the best practice and on
the other hand also allows to test both
ways: FLUIDTEMPLATE and PAGEVIEW in one
extension.

Required structural changes also showcase
what has to be done when migrating extensions
to site sets and PAGEVIEW.

Resolves: #103733
Releases: main
Change-Id: I71bb28a6e6d39b1232d5283e548044b421d3b5c9
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84037
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jörg Bösche <typo3@joergboesche.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>