[DOCS] Add sprintf() reference to f:translate ViewHelper documentation
The ViewHelper documentation for f:translate mentioned
sprintf() from PHP, but did not link to it.

This patch adds a link and also optimizes markup, plus a few more
sprintf() examples with ordered parameters and number formatting.

See forge issue for screenshots how this looks before and
after the change.

Resolves: #103808
Releases: main, 12.4
Change-Id: I237d4b668291ad68bf083e35c0bdf6510158b38b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84193
Tested-by: core-ci <typo3@b13.com>
Tested-by: Lina Wolf <112@linawolf.de>
Reviewed-by: Lina Wolf <112@linawolf.de>