[BUGFIX] Add related visibility options for PageRepository cacheIdentifier
Since #102793 the PageRepository respects `includeScheduledRecords`
from the context `VisibilityAspect` setting in the `PageRepository`
and added a runtime cache  handling in the `init()` call. However,
the cache identifier does not respect the same visibilty options
for the cache identifier like it is used for preparing the property
`where_hid_del` - and thus leads to invalid and unexpected record
retrievement, depending on context switches in a request.

This change extends the cache identifier to respect the scheduled
record setting and removes the invalid page table except comment.

Resolves: #103831
Related: #102793
Releases: main, 12.4
Change-Id: I2e7f038a04d957751bba1181c3d5324c6e0674b5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84235
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Stefan Bürk <stefan@buerk.tech>