[BUGFIX] Prevent creation of orphaned records in copy process
When records are copied, DataHandler copies its localizations as
well. This is good, but can lead to "orphan" localization records
if the target page a record is copied to is not localized to that
language.

This is inconsistent and leads to various headaches with access
rights and when localized records are synchronized later.
Also, such localized records suddenly appear, when the parent
page is localized to the missing target language, which is an
unexpected behavior for BE editors.

The patch changes this by skipping copying localizations of
records when the target page is not localized to that language.

Note DH method copyL10nOverlayRecords() will receive another
clean up with another patch to rule out some shenanigans we
stumbled upon when implementing this solution.

Resolves: #103828
Releases: main
Change-Id: Iee5530a5c6dc29180c423301b07f8cf7ecf84fc1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84237
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>