[DOCS] Improve "_assets" composer directory migration issues
It has come up a few times that users face problems how to properly
utilize the new `_assets/` hashed symlinks for `Resources/Public`
assets of extensions.

The current ChangeLog entry does not mention key searchwords
like `_assets` or `Resources/Public`. Also, vital information is
only available on external blog articles.

This patch tries to provide a central entry point and may
also be ported to the Core-API documentation once the mentioned
concepts are approved by the core-team.

ChangeLog entries are the vital source of truth for migration
instructions and basis for the documentation, thus this is the
first place to be addressed.

Resolves: #103845
Releases: main, 12.4
Change-Id: Iadeae5de624027b5303a4240b36ca501ae29a31a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84282
Tested-by: Chris Müller <typo3@brotkrueml.dev>
Tested-by: André Buchmann <andy.schliesser@gmail.com>
Reviewed-by: André Buchmann <andy.schliesser@gmail.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Chris Müller <typo3@brotkrueml.dev>