[BUGFIX] Catch errors if ini_set is not allowed
If the function ini_set is part of the setting
`disable_functions` the errors must be caught to
allow a running setup.

Most of the session settings of the install tool are
best practice and it keeps working without them. Two
of them are not: session.cookie_secure keeps the
session cookie off plain HTTP, session.cookie_httponly
keeps it away from JavaScript. Both are read back with
ini_get now, and an error is logged if they are not in
effect, so a silently degraded session does not go
unnoticed.

See https://www.php.net/manual/en/ini.core.php#ini.disable-functions

Resolves: #103842
Releases: main, 14.3
Change-Id: I5c6a88e4470a59d0a843bf29d494f41b1e6fce79
Signed-off-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/95573
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>