[TASK] Remove todo in extbase persistence backend
This change removes a 5 years old todo in extbase persistence backend.
Although the affected feature `newRecordStoragePid` currently is not
documented, it is used in various TER extensions and most likely
also in custom projects.

In relation to this change, a PR for the TYPO3 documentation has
been created on GitHub, so the setting will be documented.

Resolves: #103895
Releases: main, 12.4
Change-Id: I5c5cf7e87fdb50ab08e409e69aff2b373d39a022
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84381
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>