[FEATURE] Introduce `lint:yaml` command
The `symfony/yaml` package is a hard dependency of the
TYPO3 Core and ships a dedicated `yaml-lint` binary as
entry point for the shipped `LintCommand` class.

Most TYPO3 integrators and developers are not aware of
that dedicated binary installed along the way and only
know about the TYPO3 binary for command execution. As
the TYPO3 core relies on YAML files and parsing with an
increasing number of implementation and use cases it's
good to have linting of these files in available and easy to reach.

This change adds the `symfony/yaml` linting command to
the TYPO3 binary as feature with corresponding raise
of awareness, albeit being a development and pipeline
support feature only.

The new command can be used from the TYPO3 binary with:

  bin/typo3 lint:yaml
  bin/typo3 lint:yaml --help

directly in composer and non-composer installations.

Resolves: #104973
Releases: main
Change-Id: Ic65da691d34dccec2b292a946f59985ceadaf780
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84438
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Guido Schmechel <guido.schmechel@brandung.de>
Tested-by: Guido Schmechel <guido.schmechel@brandung.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>