[DOCS] Add details to "Trusted Host Pattern" installer warning message
The TYPO3 Contribution Guide lists DDEV as an easy way to get
started with a TYPO3-Core legacy-based installation.

The installation itself shows a big warning for the "Trusted Host
Pattern" mismatch when utilizing https:// for the installation,
which many people might do.

While the Contribution Guide mentions this as a sidenote,
the actual warning can be improved to better reflect the
reason and solution.

(Note that the installation can actually continue even
with this error, but may hint people at this being an
installation show-stopper, so it is now stated clearer,
that a working host pattern will be actually set and should
be adjusted after installation is complete)

Resolves: #103959
Related: #88846
Related: #86356
Releases: main, 12.4
Change-Id: I7988a7dad45b53d5b1c8314e1dfdcbc1d8feb6a5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84504
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>