[TASK] Deprecate namespaced shorthand validator usage in extbase
It is possible to use the namespaced shorthand notation in extbase
to add validators for properties or arguments. As an example,
`TYPO3.CMS.Extbase:NotEmpty` will be resolved as
`TYPO3\CMS\Extbase\Validation\Validator\NotEmptyValidator` or
`Vendor.Extension:Custom` will be resolved as
`\Vendor\MyExtension\Validation\Validator\CustomValidator`.

The namespaced shorthand notation for extbase validators has
however never been documented and is rarely used in extensions.

This change deprecates the usage of the namespaced shorthand
notation for validators in extbase.

Resolves: #103965
Releases: main
Change-Id: I787401cc3271cc54fed5c875a27cfde545b4e6dc
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84477
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>