[BUGFIX] Properly show multiple cropping coverAreas per cropVariant
When switching between cropVariants with multiple coverAreas,
all of the previous coverAreas are properly removed, only
showing those that belong to a specific cropVariant.

See details and screenshots in the related ticket.

(A new styleguide TCA element crop_9 is also added
to showcase multiple coverAreas in multiple cropVariants)

Resolves: #103389
Releases: main, 12.4
Change-Id: I7649aa0fe5bc01939326119b24dbdafd5f09a708
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84837
Tested-by: Andreas Nedbal <andy@pixelde.su>
Reviewed-by: Andreas Nedbal <andy@pixelde.su>
Tested-by: core-ci <typo3@b13.com>