[TASK] Add styleguide example for inline+collapseAll+date combo
The issue reported in #93743 seems to
not be reproducible in main, the datepicker
works even when "collapseAll=true" (and
"collapseAll=true") is set in the TCA.

To easily reproduce the scenario,
a new styleguide TCA element
"Form engine - inline expand single - inline_2"
is introduced and added as inline relation
to "Form engine - inline expand single".

The inline elements are dateTime inputs.

Resolves: #93743
Releases: main
Change-Id: I74da9da60a7b79213cae7d4b6b8c327972c2652a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84490
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>