[BUGFIX] Fix color preview in color picker component
With the upgrade of alwan to 2.x, CSS was completely overhauled. This
snuck in some issue regarding the color previews, where always either a
dark or light background was rendered, depending on user configuration.

This commit fixes the CSS to use the internal `--color` custom property,
which reflects the actual color value.

Resolves: #103995
Related: #102735
Releases: main
Change-Id: I326498995d2ddea5569f5e9a5c35b3e620f8872a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84522
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>